Kodak12/QuickDownload/requirements.txt
pyserial>=3.5
Pillow>=8.0.0